+++ index-matching.ts
@@ -1,4 +1,4 @@
-export { default as ReverseIndex } from './src/reverse-index';
+export { default as ReverseIndex } from './src/engine/reverse-index';
 export {default as matchNetworkFilter } from './src/matching/network';
 export { parseNetworkFilter } from './src/parsing/network-filter';
 export { mkRequest } from './src/request/interface';