+++ index.ts
@@ -1,12 +1,11 @@
 export { default as CosmeticsInjection } from './src/cosmetics-injection';
-export { default as FiltersEngine } from './src/filters-engine';
+export { default as FiltersEngine } from './src/engine/engine';
+export { default as ReverseIndex } from './src/engine/reverse-index';
 export { processRawRequest } from './src/request/raw';
 export { deserializeEngine } from './src/serialization';
 
 export {default as matchCosmeticFilter } from './src/matching/cosmetics';
 export {default as matchNetworkFilter } from './src/matching/network';
-
-export { default as ReverseIndex } from './src/reverse-index';
 
 export { parseCosmeticFilter } from './src/parsing/cosmetic-filter';
 export { parseNetworkFilter } from './src/parsing/network-filter';